requirements.txt
langchain>=0.3.0,<0.4
langgraph>=0.2.0,<0.3
streamlit>=1.40.0,<2.0
google-generativeai>=0.8.0,<1.0
anthropic>=0.40.0,<1.0
beautifulsoup4>=4.12.0,<5.0
tavily-python>=0.5.0,<1.0
python-dotenv>=1.0.0,<2.0
pandas>=2.2.0,<3.0
requests>=2.32.0,<3.0
html2text>=2024.2.26
unstructured>=0.16.0
pypdf>=5.0.0,<6.0
langchain-google-genai>=2.0.0,<3.0
langchain-anthropic>=0.3.0,<0.4
langchain-community>=0.3.0,<0.4
openai>=1.50.0,<2.0
python-docx>=1.1.0,<2.0
